+++ contracts/modules/liquidity/balancer/LiquidityModule.sol
@@ -319,8 +319,8 @@ contract LiquidityModule_Balancer is ModuleBaseWithFee {
                         _dealId,
                         _lpToken,
                         payout, // amount
-                        _la.pathTo[k * 4 + 2], // start
-                        _la.pathTo[k * 4 + 3] // end
+                        uint32(_la.pathTo[k * 4 + 2]), // start
+                        uint32(_la.pathTo[k * 4 + 3]) // end
                     );
             }
